Fix MySQL, handle SQLite multi-arg variant
